src/requirements.txt
requests
python-dateutil
suds-jurko
lxml
beautifulsoup4
